# api_swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a Travel, Renting and Ride management server, with the multiple services provided by this server.",
    "version" : "1.0.0",
    "title" : "TPM Project",
    "contact" : {
      "email" : "mathieuperez3@gmail.com"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/cdp/tmp-pw/1.0.0",
...
  "schemes" : [ "https", "http" ],
  "paths" : {
    "/api/users/" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Create user",
        "description" : "This can only be done by the user who's signing up.",
        "operationId" : "createUser",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Created user object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation"
          },
          "401" : {
            "description" : "Register failed"
          },
          "409" : {
            "description" : "There is already a user with this login or email"
          },
          "422" : {
            "description" : "Missing Arguments"
          },
          "500" : {
            "description" : "There was a problem with the database while checking if the login or email already exists"
          }
        }
      }
    },
    "/api/users/token" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Authenticate user",
        "description" : "Logs user into the system",
        "operationId" : "loginUser",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "login",
          "in" : "query",
          "description" : "The user login",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "The password in clear text",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Authentification succeeded"
          },
          "400" : {
            "description" : "Authentification failed wrong login or password"
          },
          "422" : {
            "description" : "Missing Arguments"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "User" : {
      "type" : "object",
      "properties" : {
        "login" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "role" : {
          "type" : "string"
        }
      },
      "example" : {
        "password" : "password",
        "role" : "role",
        "login" : "login",
        "email" : "email"
      }
    },
...
  }
}